PYTHON/PycharmProjects/Project/FF_5/requirements.txt
mysql-connector-python
questionary

